# docs/phase1_openapi.yaml
openapi: 3.0.3
info:
  title: 智享营销系统第 1 阶段 API
  version: 1.0.0
  description: >
    第 1 阶段 MVP 接口定义，覆盖后台管理、门店端、小程序端、微信支付、销售单微信分享收款。
    价格、库存、订单状态、客户身份、收款状态均由后端裁剪和流转，前端不得自行判断。
servers:
  - url: https://api.example.com
    description: 生产环境占位
  - url: http://localhost:8080
    description: 本地开发环境
...
security:
  - bearerAuth: []
paths:
...
  /api/store/inventory:
    get:
      tags: [StoreOrder]
      summary: 门店库存查询
      parameters:
        - name: keyword
          in: query
          description: 商品名称、SKU编码、条码
          schema:
            type: string
        - name: storeId
          in: query
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 库存列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/InventoryItem'
  /api/store/orders/{orderNo}/accept:
    post:
      tags: [StoreOrder]
      summary: 门店接单
      parameters:
        - $ref: '#/components/parameters/OrderNo'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderActionRequest'
      responses:
        '200':
          description: 接单成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/store/orders/{orderNo}/complete:
    post:
      tags: [StoreOrder]
      summary: 门店完成订单
      parameters:
        - $ref: '#/components/parameters/OrderNo'
      responses:
        '200':
          description: 完成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/store/sale-bills:
    post:
      tags: [StoreSaleBill]
      summary: 创建门店销售单
      description: 创建线下销售单或批发销售单，后端计算金额、库存、客户身份和收款状态。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaleBillCreateRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SaleBill'
    get:
      tags: [StoreSaleBill]
      summary: 销售单列表
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: collectionStatus
          in: query
          schema:
            type: string
            enum: [UNPAID, PENDING, SHARED, PARTIAL, PAID, OVERDUE, CLOSED]
        - name: keyword
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 销售单列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaleBillPageResponse'
  /api/store/sale-bills/{billNo}:
    get:
      tags: [StoreSaleBill]
      summary: 销售单详情
      parameters:
        - $ref: '#/components/parameters/BillNo'
      responses:
        '200':
          description: 销售单详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SaleBill'
  /api/store/sale-bills/{billNo}/collection-link:
    post:
      tags: [StoreSaleBill]
      summary: 生成销售单分享收款
      description: 生成微信可打开的收款页面、小程序卡片、图片或二维码。生成后本次收款金额锁定。
      parameters:
        - $ref: '#/components/parameters/BillNo'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCollectionLinkRequest'
      responses:
        '200':
          description: 生成成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CollectionLink'
  /api/store/sale-bills/{billNo}/offline-payment:
    post:
      tags: [StoreSaleBill]
      summary: 确认线下收款
      description: 现金、转账、外部微信收款后由授权人员确认，必须写操作日志。
      parameters:
        - $ref: '#/components/parameters/BillNo'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OfflinePaymentRequest'
      responses:
        '200':
          description: 确认成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /api/share/collections/{token}:
    get:
      tags: [ShareCollection]
      summary: 客户查看分享收款单
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 分享收款详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ShareCollectionDetail'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    miniappAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    OrderNo:
      name: orderNo
      in: path
      required: true
      schema:
        type: string
    BillNo:
      name: billNo
      in: path
      required: true
      schema:
        type: string
  schemas:
    ApiResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "0"
        message:
          type: string
          example: 成功
        traceId:
          type: string
          example: 01HZTRACEID
    PageResponse:
      type: object
      required: [total, page, pageSize, records]
      properties:
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        records:
          type: array
          items:
            type: object
...
    InventoryItem:
      type: object
      properties:
        storeId:
          type: integer
          format: int64
        skuId:
          type: integer
          format: int64
        skuName:
          type: string
        stockType:
          type: string
          enum: [ONLINE, OFFLINE]
        physicalQty:
          type: integer
        lockedQty:
          type: integer
        availableQty:
          type: integer
...
    OrderActionRequest:
      type: object
      properties:
        remark:
          type: string
        idempotencyKey:
          type: string
    SaleBillCreateRequest:
      type: object
      required: [storeId, items]
      properties:
        storeId:
          type: integer
          format: int64
        customerId:
          type: integer
          format: int64
          nullable: true
        customerName:
          type: string
        customerMobile:
          type: string
        discountAmount:
          type: number
          default: 0
        roundingAmount:
          type: number
          default: 0
        remark:
          type: string
        internalRemark:
          type: string
        idempotencyKey:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/SaleBillItemCreateRequest'
    SaleBillItemCreateRequest:
      type: object
      required: [skuId, totalBottleQty]
      properties:
        skuId:
          type: integer
          format: int64
        boxQty:
          type: integer
          default: 0
        bottleQty:
          type: integer
          default: 0
        totalBottleQty:
          type: integer
        unitPrice:
          type: number
          nullable: true
          description: 如为空，后端按客户身份和门店价格计算。
        priceType:
          type: string
          enum: [RETAIL, WHOLESALE, STORE]
          nullable: true
    SaleBill:
      type: object
      properties:
        billNo:
          type: string
        storeId:
          type: integer
          format: int64
        customerId:
          type: integer
          format: int64
          nullable: true
        customerName:
          type: string
        customerType:
          type: string
          enum: [RETAIL, WHOLESALE]
        businessStatus:
          type: string
          enum: [DRAFT, CREATED, COMPLETED, VOIDED, RETURNED]
        collectionStatus:
          type: string
          enum: [UNPAID, PENDING, SHARED, PARTIAL, PAID, OVERDUE, CLOSED]
        receivableAmount:
          type: number
        receivedAmount:
          type: number
        unreceivedAmount:
          type: number
        items:
          type: array
          items:
            $ref: '#/components/schemas/SaleBillItem'
    SaleBillItem:
      type: object
      properties:
        skuId:
          type: integer
          format: int64
        skuName:
          type: string
        boxQty:
          type: integer
        bottleQty:
          type: integer
        totalBottleQty:
          type: integer
        unitPrice:
          type: number
        subtotalAmount:
          type: number
    SaleBillPageResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/PageResponse'
                - type: object
                  properties:
                    records:
                      type: array
                      items:
                        $ref: '#/components/schemas/SaleBill'
    CreateCollectionLinkRequest:
      type: object
      required: [shareChannel, amount]
      properties:
        shareChannel:
          type: string
          enum: [MINIAPP_CARD, LINK, IMAGE, QR_CODE]
        amount:
          type: number
          description: 本次收款金额，不得超过销售单未收金额。
        expireHours:
          type: integer
          default: 72
        remark:
          type: string
    CollectionLink:
      type: object
      properties:
        linkNo:
          type: string
        sourceType:
          type: string
        sourceNo:
          type: string
        amount:
          type: number
        paidAmount:
          type: number
        status:
          type: string
          enum: [PENDING, PARTIAL, PAID, EXPIRED, CLOSED]
        shareUrl:
          type: string
        qrCodeUrl:
          type: string
        expireAt:
          type: string
          format: date-time
    OfflinePaymentRequest:
      type: object
      required: [amount, paymentMethod]
      properties:
        amount:
          type: number
        paymentMethod:
          type: string
          enum: [CASH, TRANSFER, OTHER_WECHAT, ALIPAY]
        remark:
          type: string
        idempotencyKey:
          type: string
...
    ShareCollectionDetail:
      type: object
      properties:
        linkNo:
          type: string
        storeName:
          type: string
        sourceNo:
          type: string
        amount:
          type: number
        paidAmount:
          type: number
        status:
          type: string
        expireAt:
          type: string
          format: date-time
        customerName:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/SaleBillItem'
...